Change approach to use filters, add sort by votes
The original UpDownUpDown required inserting a call into the Wordpress
code. This change requires no changes to Wordpress but instead uses the
hook, 'comment_text', to insert the updown vote block into the comment
stream.

This modification also causes the comments to be sorted by the net
number of upvotes.